refactor: move reservations -> spa
