+++ apps/admin-ui/src/App.tsx
@@ -9,7 +9,7 @@ import { GlobalContext } from "@context/GlobalContexts";
 import { prefixes } from "@common/urls";
 import { AuthorisationChecker } from "@component/AuthorisationChecker";
 import { Error404 } from "@component/error";
-import ReservationsRouter from "@component/reservations/ReservationRouter";
+import ReservationsRouter from "@/spa/reservations/ReservationRouter";
 import { MyUnitsRouter } from "@/spa/my-units/MyUnitsRouter";
 import NotificationsRouter from "@/spa/notifications/router";
 import ApplicationRound from "@/spa/recurring-reservations/application-rounds/[id]";
