+++ apps/admin-ui/src/common/fragments.ts
