+++ apps/admin-ui/src/hooks/reservations/index.ts
@@ -17,10 +17,11 @@ import {
   RESERVATIONS_BY_RESERVATIONUNIT,
   SINGLE_RESERVATION_QUERY,
 } from "./queries";
-import { useNotification } from "../../../../context/NotificationContext";
-import { RESERVATION_DENY_REASONS } from "../queries";
-import { OptionType } from "../../../../common/types";
-import { GQL_MAX_RESULTS_PER_QUERY } from "../../../../common/const";
+import { useNotification } from "@context/NotificationContext";
+import { OptionType } from "@common/types";
+import { GQL_MAX_RESULTS_PER_QUERY } from "@common/const";
+// TODO move to common queries
+import { RESERVATION_DENY_REASONS } from "@/spa/reservations/requested/queries";
 
 export { default as useCheckCollisions } from "./useCheckCollisions";
