+++ apps/admin-ui/src/hooks/reservations/useCheckCollisions.ts
