+++ apps/admin-ui/src/spa/my-units/[id]/recurring/completed/index.tsx
@@ -6,7 +6,8 @@ import { Container, Button } from "hds-react";
 import { ErrorBoundary } from "react-error-boundary";
 import { H1, H6 } from "common/src/common/typography";
 import { breakpoints } from "common/src/common/style";
-import { RecurringReservationsView } from "@component/reservations/requested/RecurringReservationsView";
+// TODO move to common components
+import { RecurringReservationsView } from "@/spa/reservations/requested/RecurringReservationsView";
 import { ActionsWrapper } from "@component/my-units/MyUnitRecurringReservation/commonStyling";
 import ReservationList from "@component/ReservationsList";
 import { RecurringReservationDoneParamsSchema } from "@component/my-units/schemas";
