+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/[id]/allocation/ApplicationEventCard.tsx
@@ -8,9 +8,8 @@ import type {
 } from "common/types/gql-types";
 import { SemiBold, type ReservationUnitNode, fontMedium } from "common";
 import { publicUrl } from "@/common/const";
-import { formatDuration } from "@/common/util";
+import { formatDuration, ageGroup } from "@/common/util";
 import { getApplicantName } from "@/modules/application";
-import { ageGroup } from "@/component/reservations/requested/util";
 import { filterNonNullable } from "common/src/helpers";
 import { formatTime } from "./modules/applicationRoundAllocation";
