+++ apps/admin-ui/src/spa/reservations/AllReservations.tsx
@@ -2,9 +2,9 @@ import { debounce } from "lodash";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { H1 } from "common/src/common/typography";
-import Filters, { FilterArguments, emptyState } from "./Filters";
-import ReservationsDataLoader, { Sort } from "./ReservationsDataLoader";
-import BreadcrumbWrapper from "../BreadcrumbWrapper";
+import Filters, { type FilterArguments, emptyState } from "./Filters";
+import ReservationsDataLoader, { type Sort } from "./ReservationsDataLoader";
+import BreadcrumbWrapper from "@/component/BreadcrumbWrapper";
 import { HR } from "@/component/Table";
 import { Container } from "@/styles/layout";
 import { DATE_FORMAT, formatDate } from "@/common/util";
