+++ apps/admin-ui/src/spa/reservations/requested/Calendar.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from "react-i18next";
 import { type ReservationType } from "common/types/gql-types";
 import { useModal } from "app/context/ModalContext";
 import eventStyleGetter, { legend } from "./eventStyleGetter";
-import Legend from "./Legend";
+import { Legend } from "@component/Legend";
 import EditTimeModal from "../EditTimeModal";
 import { isPossibleToEdit } from "./reservationModificationRules";
